+++ NEWS.md
@@ -1,8 +1,13 @@
 
 # renv 0.18.0  (UNRELEASED)
 
+
 * `update()` can now update packages installed from GitLab (#136) and 
   BitBucket (#1194).
+
+* `snapshot()` now standardises pak metadata so CRAN packages installed via
+  pak look the same as CRAN packages installed with renv or `install.packages()`
+  (#1239).
 
 * `install()` now keeps source when installing packages from source (#522).